Updates after review
