+++ src/main/java/org/sagebionetworks/bridge/models/activities/StudyActivityEventMap.java
@@ -7,6 +7,10 @@ import java.util.Map;
 import org.sagebionetworks.bridge.models.schedules2.StudyBurst;
 import org.sagebionetworks.bridge.models.studies.StudyCustomEvent;
 
+/**
+ * Some events include a user-defined identifier (custom events, study burst events). This class
+ * provides the custom values so we can validate an event string is using a known identifier.  
+ */
 public class StudyActivityEventMap {
 
     Map<String, ActivityEventUpdateType> customEvents;
