+++ src/test/java/org/sagebionetworks/bridge/models/activities/StudyActivityEventTest.java
@@ -3,9 +3,14 @@ package org.sagebionetworks.bridge.models.activities;
 import static org.sagebionetworks.bridge.TestConstants.CREATED_ON;
 import static org.sagebionetworks.bridge.TestConstants.MODIFIED_ON;
 import static org.sagebionetworks.bridge.TestConstants.TEST_APP_ID;
+import static org.sagebionetworks.bridge.TestConstants.TEST_STUDY_ID;
 import static org.sagebionetworks.bridge.TestConstants.TEST_USER_ID;
+import static org.sagebionetworks.bridge.models.activities.ActivityEventObjectType.QUESTION;
+import static org.sagebionetworks.bridge.models.activities.ActivityEventType.ANSWERED;
 import static org.sagebionetworks.bridge.models.activities.ActivityEventUpdateType.FUTURE_ONLY;
+import static org.sagebionetworks.bridge.models.activities.ActivityEventUpdateType.MUTABLE;
 import static org.testng.Assert.assertEquals;
+import static org.testng.Assert.assertNull;
 
 import com.fasterxml.jackson.databind.JsonNode;
 
@@ -42,5 +47,70 @@ public class StudyActivityEventTest {
         assertEquals(node.get("recordCount").intValue(), 10);
         assertEquals(node.get("updateType").textValue(), "future_only");
         assertEquals(node.get("type").textValue(), "ActivityEvent");
+    }
+    
+    
+    @Test
+    public void createsValidEvent() {
+        StudyActivityEvent event = new StudyActivityEvent.Builder()
+                .withAppId(TEST_APP_ID)
+                .withUserId(TEST_USER_ID)
+                .withStudyId(TEST_STUDY_ID)
+                .withClientTimeZone("America/Los_Angeles")
+                .withObjectType(QUESTION)
+                .withObjectId("foo")
+                .withAnswerValue("anAnswer")
+                .withTimestamp(MODIFIED_ON)
+                .withCreatedOn(CREATED_ON)
+                .withUpdateType(MUTABLE)
+                .withEventType(ANSWERED).build();
+        
+        assertEquals(event.getAppId(), TEST_APP_ID);
+        assertEquals(event.getUserId(), TEST_USER_ID);
+        assertEquals(event.getStudyId(), TEST_STUDY_ID);
+        assertEquals(event.getEventId(), "question:foo:answered=anAnswer");
+        assertEquals(event.getTimestamp(), MODIFIED_ON);
+        assertEquals(event.getAnswerValue(), "anAnswer");
+        assertEquals(event.getClientTimeZone(), "America/Los_Angeles");
+        assertEquals(event.getCreatedOn(), CREATED_ON);
+        assertEquals(event.getRecordCount(), 0);
+        assertEquals(event.getUpdateType(), MUTABLE);
+    }
+    
+    @Test
+    public void builder_nullSafe() {
+        StudyActivityEvent event = new StudyActivityEvent.Builder().build();
+        assertNull(event.getEventId());
+    }
+    
+    @Test
+    public void builder_objectTypeIsNull() {
+        StudyActivityEvent event = new StudyActivityEvent.Builder()
+                .withAppId(TEST_APP_ID)
+                .withUserId(TEST_USER_ID)
+                .withStudyId(TEST_STUDY_ID)
+                .withClientTimeZone("America/Los_Angeles")
+                .withObjectId("foo")
+                .withAnswerValue("anAnswer")
+                .withTimestamp(MODIFIED_ON)
+                .withCreatedOn(CREATED_ON)
+                .withEventType(ANSWERED).build();
+        assertNull(event.getEventId());
+    }
+    
+    @Test
+    public void builder_updateTypeIsNull() {
+        StudyActivityEvent event = new StudyActivityEvent.Builder()
+                .withAppId(TEST_APP_ID)
+                .withUserId(TEST_USER_ID)
+                .withStudyId(TEST_STUDY_ID)
+                .withClientTimeZone("America/Los_Angeles")
+                .withObjectType(QUESTION)
+                .withObjectId("foo")
+                .withAnswerValue("anAnswer")
+                .withTimestamp(MODIFIED_ON)
+                .withCreatedOn(CREATED_ON)
+                .withEventType(ANSWERED).build();
+        assertEquals(event.getUpdateType(), FUTURE_ONLY);
     }
 }
